+++ binary_playground.h
@@ -5,15 +5,15 @@
 
 typedef struct {
     int x;
-    float y;
+    double y;
     double z;
 } MyStruct;
 
-/* Sum three scalar floats and return the result */
-float sum_scalars(float a, float b, float c);
+/* Sum three scalar doubles and return the result */
+double sum_scalars(double a, double b, double c);
 
 /* In-place sum of three vectors: result stored into v1. `len` is the length. */
-void sum_vectors(float *v1, const float *v2, const float *v3, size_t len);
+void sum_vectors(double *v1, const double *v2, const double *v3, size_t len);
 
 /* Sum fields of two structs and return the resulting struct */
 MyStruct sum_structs(const MyStruct *s1, const MyStruct *s2);
